Increase Efficiency: Use MPPT mode more often
The dcdc_control now better controls which mode (boost vs. buck) is being
used when: If configured in Buck mode, boost mode is never used.

The _dcdc_control function now "knows" if it is in boost or buck mode (dcdc_dir)

The no load condition now uses the dcdc transfer power as measure, not the out current
which is not reflecting the "no dcdc load" condition correctly if a load is connected