Improve full charge detection and topping control
The battery is only full if the topping voltage is still reached AND the
current is below cut-off threshold at the same time. This was not always
detected correctly with previous 2s target voltage monitoring interval.

In addition to that, a new timer was introduced to count seconds at target
voltage and improve full charge detection due to timeout.