PWM switch: Improve overvoltage protection
In situations with full battery and sufficient solar panel power, the
battery could reach a high voltage directly after start of the PWM switch
and the control algorithm was not fast enough to limit the voltage (it
takes up to 18s to reach 5% duty cycle after starting at 95%).

This commit increases the step size in the direction of reduced power
and introduces an overvoltage flag that results in restart with 5% duty
cycle if the previous stop was caused by an overvoltage event.