Fix hard faults in EEPROM CRC calculation
The hardware CRC calculation fails hard if a non-aligned memory pointer
is passed to it. This was leading to random issues when changing
unrelated parts in the firmware so that the compiler shuffled memory
locations.

Telling GCC that it should word-align the buffer fixes the issue.

Also renaming _lock to eeprom_buf_lock.