Bat charger: Debounce recharge from trickle state
Reduce the recharge voltage threshold and consider it for transition
between trickle and bulk charging such that the trickle state is
maintained longer and batteries don't get recharged multiple times
per day.

This is also necessary for nanogrid interface control, which can try to
maintain a voltage between recharge and trickle voltage by exporting
excess solar energy towards the grid. If the battery voltage is maintained
at the trickle voltage, the solar input power might be reduced and not all
available power can be used.