ThingSet: Move info/Timestamp_s to output/Uptime_s
This makes clear that the timestamp is not the actual time, but the
seconds since the device started.

The timestamp is also not stored in NVM anymore, as the DayCount provides
information about the operating time of the device already. This way
it can be found out if there was a reset of the device based on its uptime.