Major update of ThingSet Data Object structure for v0.5
The structure has been changed to match the new ThingSet v0.5 spec which
is more suitable to be used in mobile phone apps because the data is
structured based on logical subsystems (Battery, Load) and not on data type
(conf, input, etc.) anymore.

This required some updates in the ThingSet library aswell. Especially the
update mechanism for configuration data had to be changed.

The data_objects_parser script is now able to extract the units itself,
reducing the commenting effort.

All IDs have been kept the same, so previously stored settings in charge
controllers NVM should be compatible.